+++ pkgs/cupertino_http/example/integration_test/url_session_task_test.dart
@@ -9,8 +9,151 @@ import 'package:flutter/foundation.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:test/test.dart';
 
-void testURLSessionTask(
-    URLSessionTask Function(URLSession session, Uri url) f) {
+void testWebSocketTask() {
+  group('websocket', () {
+    late HttpServer server;
+    int? lastCloseCode;
+    String? lastCloseReason;
+
+    setUp(() async {
+      lastCloseCode = null;
+      lastCloseReason = null;
+      server = await HttpServer.bind('localhost', 0)
+        ..listen((request) {
+          if (request.uri.path.endsWith('error')) {
+            request.response.statusCode = 500;
+            request.response.close();
+          } else {
+            WebSocketTransformer.upgrade(request)
+                .then((websocket) => websocket.listen((event) {
+                      final code = request.uri.queryParameters['code'];
+                      final reason = request.uri.queryParameters['reason'];
+
+                      websocket.add(event);
+                      if (!request.uri.queryParameters.containsKey('noclose')) {
+                        websocket.close(
+                            code == null ? null : int.parse(code), reason);
+                      }
+                    }, onDone: () {
+                      lastCloseCode = websocket.closeCode;
+                      lastCloseReason = websocket.closeReason;
+                    }));
+          }
+        });
+    });
+
+    tearDown(() async {
+      await server.close();
+    });
+
+    test('background session', () {
+      final session = URLSession.sessionWithConfiguration(
+          URLSessionConfiguration.backgroundSession('background'));
+      expect(
+          () => session.webSocketTaskWithRequest(URLRequest.fromUrl(
+              Uri.parse('ws://localhost:${server.port}/?noclose'))),
+          throwsUnsupportedError);
+    });
+
+    test('client code and reason', () async {
+      final session = URLSession.sharedSession();
+      final task = session.webSocketTaskWithRequest(URLRequest.fromUrl(
+          Uri.parse('ws://localhost:${server.port}/?noclose')))
+        ..resume();
+      await task
+          .sendMessage(URLSessionWebSocketMessage.fromString('Hello World!'));
+      await task.receiveMessage();
+      task.cancelWithCloseCode(
+          4998, Data.fromUint8List(Uint8List.fromList('Bye'.codeUnits)));
+
+      // Allow the server to run and save the close code.
+      while (lastCloseCode == null) {
+        await Future<void>.delayed(const Duration(milliseconds: 10));
+      }
+      expect(lastCloseCode, 4998);
+      expect(lastCloseReason, 'Bye');
+    });
+
+    test('server code and reason', () async {
+      final session = URLSession.sharedSession();
+      final task = session.webSocketTaskWithRequest(URLRequest.fromUrl(
+          Uri.parse('ws://localhost:${server.port}/?code=4999&reason=fun')))
+        ..resume();
+      await task
+          .sendMessage(URLSessionWebSocketMessage.fromString('Hello World!'));
+      await task.receiveMessage();
+      await expectLater(task.receiveMessage(),
+          throwsA(isA<Error>().having((e) => e.code, 'code', 57 // NOT_CONNECTED
+              )));
+
+      expect(task.closeCode, 4999);
+      expect(task.closeReason!.bytes, 'fun'.codeUnits);
+      task.cancel();
+    });
+
+    test('data message', () async {
+      final session = URLSession.sharedSession();
+      final task = session.webSocketTaskWithRequest(
+          URLRequest.fromUrl(Uri.parse('ws://localhost:${server.port}')))
+        ..resume();
+      await task.sendMessage(URLSessionWebSocketMessage.fromData(
+          Data.fromUint8List(Uint8List.fromList([1, 2, 3]))));
+      final receivedMessage = await task.receiveMessage();
+      expect(receivedMessage.type,
+          URLSessionWebSocketMessageType.urlSessionWebSocketMessageTypeData);
+      expect(receivedMessage.data!.bytes, [1, 2, 3]);
+      expect(receivedMessage.string, null);
+      task.cancel();
+    });
+
+    test('text message', () async {
+      final session = URLSession.sharedSession();
+      final task = session.webSocketTaskWithRequest(
+          URLRequest.fromUrl(Uri.parse('ws://localhost:${server.port}')))
+        ..resume();
+      await task
+          .sendMessage(URLSessionWebSocketMessage.fromString('Hello World!'));
+      final receivedMessage = await task.receiveMessage();
+      expect(receivedMessage.type,
+          URLSessionWebSocketMessageType.urlSessionWebSocketMessageTypeString);
+      expect(receivedMessage.data, null);
+      expect(receivedMessage.string, 'Hello World!');
+      task.cancel();
+    });
+
+    test('send failure', () async {
+      final session = URLSession.sharedSession();
+      final task = session.webSocketTaskWithRequest(
+          URLRequest.fromUrl(Uri.parse('ws://localhost:${server.port}/error')))
+        ..resume();
+      await expectLater(
+          task.sendMessage(
+              URLSessionWebSocketMessage.fromString('Hello World!')),
+          throwsA(isA<Error>().having(
+              (e) => e.code, 'code', -1011 // NSURLErrorBadServerResponse
+              )));
+      task.cancel();
+    });
+
+    test('receive failure', () async {
+      final session = URLSession.sharedSession();
+      final task = session.webSocketTaskWithRequest(
+          URLRequest.fromUrl(Uri.parse('ws://localhost:${server.port}')))
+        ..resume();
+      await task
+          .sendMessage(URLSessionWebSocketMessage.fromString('Hello World!'));
+      await task.receiveMessage();
+      await expectLater(task.receiveMessage(),
+          throwsA(isA<Error>().having((e) => e.code, 'code', 57 // NOT_CONNECTED
+              )));
+      task.cancel();
+    });
+  });
+}
+
+void testURLSessionTaskCommon(
+    URLSessionTask Function(URLSession session, Uri url) f,
+    {bool suspendedAfterCancel = false}) {
   group('task states', () {
     late HttpServer server;
     late URLSessionTask task;
@@ -44,7 +187,11 @@ void testURLSessionTask(
 
     test('cancel', () {
       task.cancel();
-      expect(task.state, URLSessionTaskState.urlSessionTaskStateCanceling);
+      if (suspendedAfterCancel) {
+        expect(task.state, URLSessionTaskState.urlSessionTaskStateSuspended);
+      } else {
+        expect(task.state, URLSessionTaskState.urlSessionTaskStateCanceling);
+      }
       expect(task.response, null);
       task.toString(); // Just verify that there is no crash.
     });
@@ -223,12 +370,21 @@ void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
   group('data task', () {
-    testURLSessionTask(
+    testURLSessionTaskCommon(
         (session, uri) => session.dataTaskWithRequest(URLRequest.fromUrl(uri)));
   });
 
   group('download task', () {
-    testURLSessionTask((session, uri) =>
+    testURLSessionTaskCommon((session, uri) =>
         session.downloadTaskWithRequest(URLRequest.fromUrl(uri)));
   });
+
+  group('websocket task', () {
+    testURLSessionTaskCommon(
+        (session, uri) =>
+            session.webSocketTaskWithRequest(URLRequest.fromUrl(uri)),
+        suspendedAfterCancel: true);
+  });
+
+  testWebSocketTask();
 }