+++ pkgs/http_client_conformance_tests/lib/src/isolate_test.dart
@@ -5,7 +5,7 @@
 import 'dart:isolate' if (dart.library.html) 'dummy_isolate.dart';
 
 import 'package:async/async.dart';
-import 'package:http/http.dart';
+import 'package:cancellation_token_http/http.dart';
 import 'package:stream_channel/stream_channel.dart';
 import 'package:test/test.dart';